Remove empty lines
