+++ cmd/oras/push.go
@@ -86,13 +86,9 @@ Example - Push file to the HTTP registry:
 
 func runPush(opts pushOptions) error {
 	ctx, _ := opts.SetLoggerLevel()
-
-	// Load annotations
-	var annotations map[string]map[string]string
-	if opts.ManifestAnnotations != "" {
-		if err := decodeJSON(opts.ManifestAnnotations, &annotations); err != nil {
-			return err
-		}
+	annotations, err := opts.LoadAnnotations()
+	if err != nil {
+		return err
 	}
 
 	// Prepare manifest
